refactor(cli): split initAction.ts into plan/ and project/ modules
Extract helper functions from the 1175-line initAction.ts into
one-function-per-file modules under plan/ and project/ subdirectories.
initAction.ts now contains only the main orchestrator, flag validation,
and auth — down to ~300 lines.
