+++ packages/@sanity/cli/src/actions/init/initAction.ts
@@ -1,42 +1,20 @@
-import path from 'node:path'
 import {styleText} from 'node:util'
 
 import {type SanityOrgUser, subdebug, type TelemetryUserProperties} from '@sanity/cli-core'
-import {confirm, input, logSymbols, select, Separator, spinner} from '@sanity/cli-core/ux'
-import {type DatasetAclMode, isHttpError} from '@sanity/client'
+import {logSymbols, spinner} from '@sanity/cli-core/ux'
 import {type TelemetryTrace} from '@sanity/telemetry'
 import {type Framework, frameworks} from '@vercel/frameworks'
 import deburr from 'lodash-es/deburr.js'
 
 import {promptForConfigFiles} from '../../prompts/init/nextjs.js'
-import {promptForDatasetName} from '../../prompts/promptForDatasetName.js'
-import {promptForDefaultConfig} from '../../prompts/promptForDefaultConfig.js'
-import {promptForOrganizationName} from '../../prompts/promptForOrganizationName.js'
-import {createDataset as createDatasetService, listDatasets} from '../../services/datasets.js'
-import {getProjectFeatures} from '../../services/getProjectFeatures.js'
-import {
-  createOrganization,
-  listOrganizations,
-  type OrganizationCreateResponse,
-  type ProjectOrganization,
-} from '../../services/organizations.js'
-import {getPlanId, getPlanIdFromCoupon} from '../../services/plans.js'
-import {createProject, listProjects} from '../../services/projects.js'
 import {getCliUser} from '../../services/user.js'
 import {CLIInitStepCompleted, type InitStepResult} from '../../telemetry/init.telemetry.js'
 import {detectFrameworkRecord} from '../../util/detectFramework.js'
-import {absolutify, validateEmptyPath} from '../../util/fsUtils.js'
 import {getProjectDefaults} from '../../util/getProjectDefaults.js'
 import {validateSession} from '../auth/ensureAuthenticated.js'
 import {getProviderName} from '../auth/getProviderName.js'
 import {login} from '../auth/login/login.js'
-import {createDataset} from '../dataset/create.js'
 import {setupMCP} from '../mcp/setupMCP.js'
-import {findOrganizationByUserName} from '../organizations/findOrganizationByUserName.js'
-import {getOrganizationChoices} from '../organizations/getOrganizationChoices.js'
-import {getOrganizationsWithAttachGrantInfo} from '../organizations/getOrganizationsWithAttachGrantInfo.js'
-import {hasProjectAttachGrant} from '../organizations/hasProjectAttachGrant.js'
-import {type OrganizationChoices} from '../organizations/types.js'
 import {checkNextJsReactCompatibility} from './checkNextJsReactCompatibility.js'
 import {determineAppTemplate} from './determineAppTemplate.js'
 import {createOrAppendEnvVars} from './env/createOrAppendEnvVars.js'
@@ -45,6 +23,10 @@ import {InitError} from './initError.js'
 import {flagOrDefault, shouldPrompt, writeStagingEnvIfNeeded} from './initHelpers.js'
 import {initNextJs} from './initNextJs.js'
 import {initStudio} from './initStudio.js'
+import {getPlan} from './plan/getPlan.js'
+import {createProjectFromName} from './project/createProjectFromName.js'
+import {getProjectDetails} from './project/getProjectDetails.js'
+import {getProjectOutputPath} from './project/getProjectOutputPath.js'
 import {checkIsRemoteTemplate, getGitHubRepoInfo, type RepoInfo} from './remoteTemplate.js'
 import {type InitContext, type InitOptions} from './types.js'
 
@@ -332,59 +314,6 @@ function checkFlagsInUnattendedMode(
   }
 }
 
-async function createProjectFromName({
-  coupon,
-  createProjectName,
-  dataset,
-  organization,
-  planId,
-  user,
-  visibility,
-}: {
-  coupon: string | undefined
-  createProjectName: string
-  dataset: string | undefined
-  organization: string | undefined
-  planId: string | undefined
-  user: SanityOrgUser
-  visibility: 'private' | 'public' | undefined
-}): Promise<string> {
-  debug('--project-name specified, creating a new project')
-
-  let orgForCreateProjectFlag = organization
-
-  if (!orgForCreateProjectFlag) {
-    debug('no organization specified, selecting one')
-    const organizations = await listOrganizations()
-    orgForCreateProjectFlag = await promptUserForOrganization({
-      organizations,
-      user,
-    })
-  }
-
-  debug('creating a new project')
-  const createdProject = await createProject({
-    displayName: createProjectName.trim(),
-    metadata: {coupon},
-    organizationId: orgForCreateProjectFlag,
-    subscription: planId ? {planId} : undefined,
-  })
-
-  debug('Project with ID %s created', createdProject.projectId)
-  if (dataset) {
-    debug('--dataset specified, creating dataset (%s)', dataset)
-    const spin = spinner('Creating dataset').start()
-    await createDatasetService({
-      aclMode: visibility as DatasetAclMode | undefined,
-      datasetName: dataset,
-      projectId: createdProject.projectId,
-    })
-    spin.succeed()
-  }
-
-  return createdProject.projectId
-}
-
 async function ensureAuthenticated(
   options: InitOptions,
   output: InitContext['output'],
@@ -425,751 +354,4 @@ async function ensureAuthenticated(
     `${logSymbols.success} You are logged in as ${loggedInUser.email} using ${getProviderName(loggedInUser.provider)}`,
   )
   return {user: loggedInUser}
-}
-
-async function getOrCreateDataset(opts: {
-  dataset?: string
-  defaultConfig: boolean | undefined
-  displayName: string
-  output: InitContext['output']
-  projectId: string
-  showDefaultConfigPrompt: boolean
-  unattended: boolean
-  visibility?: string
-}): Promise<{
-  datasetName: string
-  userAction: 'create' | 'none' | 'select'
-}> {
-  const {dataset, visibility} = opts
-  let {defaultConfig} = opts
-
-  if (dataset && opts.unattended) {
-    return {datasetName: dataset, userAction: 'none'}
-  }
-
-  const [datasets, projectFeatures] = await Promise.all([
-    listDatasets(opts.projectId),
-    getProjectFeatures(opts.projectId),
-  ])
-
-  if (dataset) {
-    debug('User has specified dataset through a flag (%s)', dataset)
-    const existing = datasets.find((ds) => ds.name === dataset)
-    if (!existing) {
-      debug('Specified dataset not found, creating it')
-      await createDataset({
-        datasetName: dataset,
-        forcePublic: defaultConfig,
-        output: opts.output,
-        projectFeatures,
-        projectId: opts.projectId,
-        visibility,
-      })
-    }
-
-    return {datasetName: dataset, userAction: 'none'}
-  }
-
-  if (opts.unattended) {
-    debug('Unattended mode without --dataset, defaulting to "production" dataset')
-    const datasetName = 'production'
-    const existing = datasets.find((ds) => ds.name === datasetName)
-    if (!existing) {
-      await createDataset({
-        datasetName,
-        forcePublic: visibility === undefined,
-        isUnattended: true,
-        output: opts.output,
-        projectFeatures,
-        projectId: opts.projectId,
-        visibility,
-      })
-    }
-    return {datasetName, userAction: existing ? 'none' : 'create'}
-  }
-
-  if (datasets.length === 0) {
-    debug('No datasets found for project, prompting for name')
-    if (opts.showDefaultConfigPrompt) {
-      defaultConfig = await promptForDefaultConfig()
-    }
-    const name = defaultConfig
-      ? 'production'
-      : await promptForDatasetName({
-          message: 'Name of your first dataset:',
-        })
-    await createDataset({
-      datasetName: name,
-      forcePublic: defaultConfig,
-      output: opts.output,
-      projectFeatures,
-      projectId: opts.projectId,
-      visibility,
-    })
-    return {datasetName: name, userAction: 'create'}
-  }
-
-  debug(`User has ${datasets.length} dataset(s) already, showing list of choices`)
-  const datasetChoices = datasets.map((dataset) => ({value: dataset.name}))
-
-  const selected = await select({
-    choices: [{name: 'Create new dataset', value: 'new'}, new Separator(), ...datasetChoices],
-    message: 'Select dataset to use',
-  })
-
-  if (selected === 'new') {
-    const existingDatasetNames = datasets.map((ds) => ds.name)
-    debug('User wants to create a new dataset, prompting for name')
-    if (opts.showDefaultConfigPrompt && !existingDatasetNames.includes('production')) {
-      defaultConfig = await promptForDefaultConfig()
-    }
-
-    const newDatasetName = defaultConfig
-      ? 'production'
-      : await promptForDatasetName(
-          {
-            message: 'Dataset name:',
-          },
-          existingDatasetNames,
-        )
-    await createDataset({
-      datasetName: newDatasetName,
-      forcePublic: defaultConfig,
-      output: opts.output,
-      projectFeatures,
-      projectId: opts.projectId,
-      visibility,
-    })
-    return {datasetName: newDatasetName, userAction: 'create'}
-  }
-
-  debug(`Returning selected dataset (${selected})`)
-  return {datasetName: selected, userAction: 'select'}
-}
-
-async function getOrCreateProject({
-  coupon,
-  newProject,
-  organization,
-  planId,
-  project,
-  unattended,
-  user,
-}: {
-  coupon?: string
-  newProject: string | undefined
-  organization: string | undefined
-  planId: string | undefined
-  project: string | undefined
-  unattended: boolean
-  user: SanityOrgUser
-}): Promise<{
-  displayName: string
-  isFirstProject: boolean
-  projectId: string
-  userAction: 'create' | 'select'
-}> {
-  const projectId = project || newProject
-  let projects
-  let organizations: ProjectOrganization[]
-
-  try {
-    const [allProjects, allOrgs] = await Promise.all([listProjects(), listOrganizations()])
-    projects = allProjects.toSorted((a, b) => b.createdAt.localeCompare(a.createdAt))
-    organizations = allOrgs
-  } catch (err: unknown) {
-    if (unattended && projectId) {
-      return {
-        displayName: 'Unknown project',
-        isFirstProject: false,
-        projectId,
-        userAction: 'select',
-      }
-    }
-    const message = err instanceof Error ? err.message : String(err)
-    throw new InitError(`Failed to communicate with the Sanity API:\n${message}`, 1)
-  }
-
-  if (projects.length === 0 && unattended) {
-    throw new InitError('No projects found for current user', 1)
-  }
-
-  if (projectId) {
-    const proj = projects.find((p) => p.id === projectId)
-    if (!proj && !unattended) {
-      throw new InitError(
-        `Given project ID (${projectId}) not found, or you do not have access to it`,
-        1,
-      )
-    }
-
-    return {
-      displayName: proj ? proj.displayName : 'Unknown project',
-      isFirstProject: false,
-      projectId,
-      userAction: 'select',
-    }
-  }
-
-  if (organization) {
-    const org =
-      organizations.find((o) => o.id === organization) ||
-      organizations.find((o) => o.slug === organization)
-
-    if (!org) {
-      throw new InitError(
-        `Given organization ID (${organization}) not found, or you do not have access to it`,
-        1,
-      )
-    }
-
-    if (!(await hasProjectAttachGrant(organization))) {
-      throw new InitError(
-        'You lack the necessary permissions to attach a project to this organization',
-        1,
-      )
-    }
-  }
-
-  const isUsersFirstProject = projects.length === 0
-  if (isUsersFirstProject || coupon) {
-    debug(
-      isUsersFirstProject
-        ? 'No projects found for user, prompting for name'
-        : 'Using a coupon - skipping project selection',
-    )
-
-    const created = await promptForProjectCreation({
-      coupon,
-      isUsersFirstProject,
-      organizationId: organization,
-      organizations,
-      planId,
-      user,
-    })
-
-    return {
-      ...created,
-      isFirstProject: isUsersFirstProject,
-      userAction: 'create',
-    }
-  }
-
-  debug(`User has ${projects.length} project(s) already, showing list of choices`)
-
-  const projectChoices = projects.map((project) => ({
-    name: `${project.displayName} (${project.id})`,
-    value: project.id,
-  }))
-
-  const selected = await select({
-    choices: [{name: 'Create new project', value: 'new'}, new Separator(), ...projectChoices],
-    message: 'Create a new project or select an existing one',
-  })
-
-  if (selected === 'new') {
-    debug('User wants to create a new project, prompting for name')
-
-    const created = await promptForProjectCreation({
-      coupon,
-      isUsersFirstProject,
-      organizationId: organization,
-      organizations,
-      planId,
-      user,
-    })
-
-    return {
-      ...created,
-      isFirstProject: isUsersFirstProject,
-      userAction: 'create',
-    }
-  }
-
-  debug(`Returning selected project (${selected})`)
-  return {
-    displayName: projects.find((proj) => proj.id === selected)?.displayName || '',
-    isFirstProject: isUsersFirstProject,
-    projectId: selected,
-    userAction: 'select',
-  }
-}
-
-async function getPlan(
-  options: InitOptions,
-  output: InitContext['output'],
-  trace: TelemetryTrace<TelemetryUserProperties, InitStepResult>,
-): Promise<string | undefined> {
-  const intendedPlan = options.projectPlan
-  const intendedCoupon = options.coupon
-
-  if (intendedCoupon) {
-    return verifyCoupon(intendedCoupon, options.unattended, output, trace)
-  } else if (intendedPlan) {
-    return verifyPlan(intendedPlan, options.unattended, output, trace)
-  } else {
-    return undefined
-  }
-}
-
-async function getProjectDetails({
-  coupon,
-  dataset,
-  datasetDefault,
-  isAppTemplate,
-  newProject,
-  organization,
-  output,
-  planId,
-  project,
-  showDefaultConfigPrompt,
-  trace,
-  unattended,
-  user,
-  visibility,
-}: {
-  coupon: string | undefined
-  dataset: string | undefined
-  datasetDefault: boolean
-  isAppTemplate: boolean
-  newProject: string | undefined
-  organization: string | undefined
-  output: InitContext['output']
-  planId: string | undefined
-  project: string | undefined
-  showDefaultConfigPrompt: boolean
-  trace: TelemetryTrace<TelemetryUserProperties, InitStepResult>
-  unattended: boolean
-  user: SanityOrgUser
-  visibility: 'private' | 'public' | undefined
-}): Promise<{
-  datasetName: string
-  displayName: string
-  isFirstProject: boolean
-  organizationId?: string
-  projectId: string
-  schemaUrl?: string
-}> {
-  if (isAppTemplate) {
-    let appOrganizationId: string | undefined = organization
-    if (!appOrganizationId) {
-      let organizations: ProjectOrganization[]
-      try {
-        organizations = await listOrganizations()
-      } catch (err: unknown) {
-        const message = err instanceof Error ? err.message : String(err)
-        throw new InitError(`Failed to communicate with the Sanity API:\n${message}`, 1)
-      }
-      appOrganizationId = await promptUserForOrganization({
-        isAppTemplate: true,
-        organizations,
-        user,
-      })
-    }
-
-    const {
-      datasetName: appDatasetName,
-      displayName: appDisplayName,
-      projectId: appProjectId,
-    } = await promptForAppTemplateSetup({
-      coupon,
-      dataset,
-      newProject,
-      organization,
-      organizationId: appOrganizationId,
-      output,
-      planId,
-      project,
-      trace,
-      unattended,
-      user,
-      visibility,
-    })
-
-    return {
-      datasetName: appDatasetName,
-      displayName: appDisplayName,
-      isFirstProject: false,
-      organizationId: appOrganizationId,
-      projectId: appProjectId,
-    }
-  }
-
-  debug('Prompting user to select or create a project')
-  const projectResult = await getOrCreateProject({
-    coupon,
-    newProject,
-    organization,
-    planId,
-    project,
-    unattended,
-    user,
-  })
-  debug(`Project with name ${projectResult.displayName} selected`)
-
-  debug('Prompting user to select or create a dataset')
-  const datasetResult = await getOrCreateDataset({
-    dataset,
-    defaultConfig: datasetDefault || undefined,
-    displayName: projectResult.displayName,
-    output,
-    projectId: projectResult.projectId,
-    showDefaultConfigPrompt,
-    unattended,
-    visibility,
-  })
-  debug(`Dataset with name ${datasetResult.datasetName} selected`)
-
-  trace.log({
-    datasetName: datasetResult.datasetName,
-    selectedOption: datasetResult.userAction,
-    step: 'createOrSelectDataset',
-    visibility,
-  })
-
-  return {
-    datasetName: datasetResult.datasetName,
-    displayName: projectResult.displayName,
-    isFirstProject: projectResult.isFirstProject,
-    projectId: projectResult.projectId,
-  }
-}
-
-async function getProjectOutputPath({
-  initFramework,
-  outputPath,
-  sluggedName,
-  unattended,
-  useEnv,
-  workDir,
-}: {
-  initFramework: boolean
-  outputPath: string | undefined
-  sluggedName: string
-  unattended: boolean
-  useEnv: boolean
-  workDir: string
-}): Promise<string> {
-  const specifiedPath = outputPath && path.resolve(outputPath)
-  if (unattended || specifiedPath || useEnv || initFramework) {
-    return specifiedPath || workDir
-  }
-
-  const inputPath = await input({
-    default: path.join(workDir, sluggedName),
-    message: 'Project output path:',
-    validate: validateEmptyPath,
-  })
-
-  return absolutify(inputPath)
-}
-
-async function promptForAppTemplateSetup({
-  coupon,
-  dataset,
-  newProject,
-  organization,
-  organizationId,
-  output,
-  planId,
-  project,
-  trace,
-  unattended,
-  user,
-  visibility,
-}: {
-  coupon: string | undefined
-  dataset: string | undefined
-  newProject: string | undefined
-  organization: string | undefined
-  organizationId: string | undefined
-  output: InitContext['output']
-  planId: string | undefined
-  project: string | undefined
-  trace: TelemetryTrace<TelemetryUserProperties, InitStepResult>
-  unattended: boolean
-  user: SanityOrgUser
-  visibility: 'private' | 'public' | undefined
-}): Promise<{datasetName: string; displayName: string; projectId: string}> {
-  if (unattended) {
-    if (!project && !newProject) {
-      return {datasetName: '', displayName: '', projectId: ''}
-    }
-    const projectResult = await getOrCreateProject({
-      coupon,
-      newProject,
-      organization,
-      planId,
-      project,
-      unattended,
-      user,
-    })
-    const datasetResult = await getOrCreateDataset({
-      dataset,
-      defaultConfig: undefined,
-      displayName: projectResult.displayName,
-      output,
-      projectId: projectResult.projectId,
-      showDefaultConfigPrompt: false,
-      unattended,
-      visibility,
-    })
-    return {
-      datasetName: datasetResult.datasetName,
-      displayName: projectResult.displayName,
-      projectId: projectResult.projectId,
-    }
-  }
-
-  const projects = (await listProjects()).toSorted((a, b) => b.createdAt.localeCompare(a.createdAt))
-
-  const projectChoices = projects.map((p) => ({
-    name: `${p.displayName} (${p.id})`,
-    value: p.id,
-  }))
-
-  const SKIP_PROJECT = '__skip__'
-  const NEW_PROJECT = '__new__'
-
-  const selected = await select({
-    choices: [
-      {name: "Skip — I'll configure later", value: SKIP_PROJECT},
-      {name: 'Create new project', value: NEW_PROJECT},
-      ...(projectChoices.length > 0 ? [new Separator(), ...projectChoices] : []),
-    ],
-    message: 'Configure a project for this app?',
-  })
-
-  if (selected === SKIP_PROJECT) {
-    trace.log({selectedOption: 'skip', step: 'configureAppProject'})
-    return {datasetName: '', displayName: '', projectId: ''}
-  }
-
-  trace.log({
-    selectedOption: selected === NEW_PROJECT ? 'create' : 'existing',
-    step: 'configureAppProject',
-  })
-
-  const projectResult =
-    selected === NEW_PROJECT
-      ? await promptForProjectCreation({
-          coupon,
-          isUsersFirstProject: projects.length === 0,
-          organizationId,
-          organizations: [],
-          planId,
-          user,
-        })
-      : {
-          displayName: projects.find((p) => p.id === selected)?.displayName ?? '',
-          projectId: selected,
-        }
-
-  const datasetResult = await getOrCreateDataset({
-    dataset,
-    defaultConfig: undefined,
-    displayName: projectResult.displayName,
-    output,
-    projectId: projectResult.projectId,
-    showDefaultConfigPrompt: false,
-    unattended: false,
-    visibility,
-  })
-  return {
-    datasetName: datasetResult.datasetName,
-    displayName: projectResult.displayName,
-    projectId: projectResult.projectId,
-  }
-}
-
-async function promptForProjectCreation({
-  coupon,
-  isUsersFirstProject,
-  organizationId,
-  organizations,
-  planId,
-  user,
-}: {
-  coupon: string | undefined
-  isUsersFirstProject: boolean
-  organizationId: string | undefined
-  organizations: ProjectOrganization[]
-  planId: string | undefined
-  user: SanityOrgUser
-}) {
-  const projectName = await input({
-    default: 'My Sanity Project',
-    message: 'Project name:',
-    validate(val) {
-      if (!val || val.trim() === '') {
-        return 'Project name cannot be empty'
-      }
-
-      if (val.length > 80) {
-        return 'Project name cannot be longer than 80 characters'
-      }
-
-      return true
-    },
-  })
-
-  const org = organizationId || (await promptUserForOrganization({organizations, user}))
-
-  const newProjectResult = await createProject({
-    displayName: projectName,
-    metadata: {coupon},
-    organizationId: org,
-    subscription: planId ? {planId} : undefined,
-  })
-
-  return {
-    ...newProjectResult,
-    isFirstProject: isUsersFirstProject,
-    userAction: 'create' as const,
-  }
-}
-
-async function promptUserForNewOrganization(
-  user: SanityOrgUser,
-): Promise<OrganizationCreateResponse> {
-  const name = await promptForOrganizationName(user)
-
-  const spin = spinner('Creating organization').start()
-  const organization = await createOrganization(name)
-  spin.succeed()
-
-  return organization
-}
-
-async function promptUserForOrganization({
-  isAppTemplate = false,
-  organizations,
-  user,
-}: {
-  isAppTemplate?: boolean
-  organizations: ProjectOrganization[]
-  user: SanityOrgUser
-}) {
-  if (organizations.length === 0) {
-    const newOrganization = await promptUserForNewOrganization(user)
-    return newOrganization.id
-  }
-
-  let organizationChoices: OrganizationChoices
-  let defaultOrganizationId: string | undefined
-
-  if (isAppTemplate) {
-    organizationChoices = getOrganizationChoices(organizations)
-    defaultOrganizationId =
-      organizations.length === 1
-        ? organizations[0].id
-        : findOrganizationByUserName(organizations, user)
-  } else {
-    debug(`User has ${organizations.length} organization(s), checking attach access`)
-    const withGrantInfo = await getOrganizationsWithAttachGrantInfo(organizations)
-    const withAttach = withGrantInfo.filter(({hasAttachGrant}) => hasAttachGrant)
-
-    debug('User has attach access to %d organizations.', withAttach.length)
-    organizationChoices = getOrganizationChoices(withGrantInfo)
-    defaultOrganizationId =
-      withAttach.length === 1
-        ? withAttach[0].organization.id
-        : findOrganizationByUserName(organizations, user)
-  }
-
-  const chosenOrg = await select({
-    choices: organizationChoices,
-    default: defaultOrganizationId || undefined,
-    message: 'Select organization:',
-  })
-
-  if (chosenOrg === '-new-') {
-    const newOrganization = await promptUserForNewOrganization(user)
-    return newOrganization.id
-  }
-
-  return chosenOrg || undefined
-}
-
-async function verifyCoupon(
-  intendedCoupon: string,
-  unattended: boolean,
-  output: InitContext['output'],
-  trace: TelemetryTrace<TelemetryUserProperties, InitStepResult>,
-): Promise<string | undefined> {
-  try {
-    const planId = await getPlanIdFromCoupon(intendedCoupon)
-    output.log(`Coupon "${intendedCoupon}" validated!\n`)
-    return planId
-  } catch (err: unknown) {
-    if (!isHttpError(err) || err.statusCode !== 404) {
-      const message = err instanceof Error ? err.message : `${err}`
-      throw new InitError(`Unable to validate coupon, please try again later:\n\n${message}`, 1)
-    }
-
-    const useDefaultPlan =
-      unattended ||
-      (await confirm({
-        default: true,
-        message: `Coupon "${intendedCoupon}" is not available, use default plan instead?`,
-      }))
-
-    if (unattended) {
-      output.warn(`Coupon "${intendedCoupon}" is not available - using default plan`)
-    }
-
-    trace.log({
-      coupon: intendedCoupon,
-      selectedOption: useDefaultPlan ? 'yes' : 'no',
-      step: 'useDefaultPlanCoupon',
-    })
-
-    if (useDefaultPlan) {
-      output.log('Using default plan.')
-      return undefined
-    }
-
-    throw new InitError(`Coupon "${intendedCoupon}" does not exist`, 1)
-  }
-}
-
-async function verifyPlan(
-  intendedPlan: string,
-  unattended: boolean,
-  output: InitContext['output'],
-  trace: TelemetryTrace<TelemetryUserProperties, InitStepResult>,
-): Promise<string | undefined> {
-  try {
-    const planId = await getPlanId(intendedPlan)
-    return planId
-  } catch (err: unknown) {
-    if (!isHttpError(err) || err.statusCode !== 404) {
-      const message = err instanceof Error ? err.message : `${err}`
-      throw new InitError(`Unable to validate plan, please try again later:\n\n${message}`, 1)
-    }
-
-    const useDefaultPlan =
-      unattended ||
-      (await confirm({
-        default: true,
-        message: `Project plan "${intendedPlan}" does not exist, use default plan instead?`,
-      }))
-
-    if (unattended) {
-      output.warn(`Project plan "${intendedPlan}" does not exist - using default plan`)
-    }
-
-    trace.log({
-      planId: intendedPlan,
-      selectedOption: useDefaultPlan ? 'yes' : 'no',
-      step: 'useDefaultPlanId',
-    })
-
-    if (useDefaultPlan) {
-      output.log('Using default plan.')
-      return undefined
-    }
-
-    throw new InitError(`Plan id "${intendedPlan}" does not exist`, 1)
-  }
 }
